+++ packages/web3-utils/src/formatter.ts
@@ -72,10 +72,25 @@ export type FormatType<T, F extends DataFormat> = number extends Extract<T, Numb
 	  }
 	: T;
 
-const findSchemaByDataPath = (schema: JsonSchema, dataPath: string[]): JsonSchema | undefined => {
+const findSchemaByDataPath = (
+	schema: JsonSchema,
+	dataPath: string[],
+	oneOfPath: [string, number][] = [],
+): JsonSchema | undefined => {
 	let result: JsonSchema = { ...schema } as JsonSchema;
+	let previousDataPath: string | undefined;
 
 	for (const dataPart of dataPath) {
+		if (result.oneOf && previousDataPath) {
+			const path = oneOfPath.find(function (element: [string, number]) {
+				return (this as unknown as string) === element[0];
+			}, previousDataPath ?? '');
+
+			if (path && path[0] === previousDataPath) {
+				// eslint-disable-next-line @typescript-eslint/no-unsafe-assignment, @typescript-eslint/no-unsafe-member-access
+				result = result.oneOf[path[1]];
+			}
+		}
 		if (!result.properties && !result.items) {
 			return undefined;
 		}
@@ -95,6 +110,8 @@ const findSchemaByDataPath = (schema: JsonSchema, dataPath: string[]): JsonSchem
 		} else if (result.items && Array.isArray(result.items)) {
 			result = (result.items as JsonSchema[])[parseInt(dataPart, 10)];
 		}
+
+		if (result && dataPart) previousDataPath = dataPart;
 	}
 
 	return result;
@@ -145,6 +162,7 @@ export const convert = (
 	schema: JsonSchema,
 	dataPath: string[],
 	format: DataFormat,
+	oneOfPath: [string, number][] = [],
 ) => {
 	// If it's a scalar value
 	if (!isObject(data) && !Array.isArray(data)) {
@@ -155,7 +173,7 @@ export const convert = (
 
 	for (const [key, value] of Object.entries(object)) {
 		dataPath.push(key);
-		const schemaProp = findSchemaByDataPath(schema, dataPath);
+		const schemaProp = findSchemaByDataPath(schema, dataPath, oneOfPath);
 
 		// If value is a scaler value
 		if (isNullish(schemaProp)) {
@@ -186,19 +204,19 @@ export const convert = (
 				// over each possible schema and check if they type of the schema
 				// matches the type of value[0], and if so we use the oneOfSchemaProp
 				// as the schema for formatting
-				for (const oneOfSchemaProp of schemaProp.oneOf) {
+				// eslint-disable-next-line @typescript-eslint/no-unsafe-member-access, @typescript-eslint/no-unsafe-call
+				schemaProp.oneOf.forEach((oneOfSchemaProp: JsonSchema, index: number) => {
 					if (
 						!Array.isArray(schemaProp?.items) &&
 						((typeof value[0] === 'object' &&
-							((oneOfSchemaProp as JsonSchema)?.items as JsonSchema)?.type ===
-								'object') ||
+							(oneOfSchemaProp?.items as JsonSchema)?.type === 'object') ||
 							(typeof value[0] === 'string' &&
-								((oneOfSchemaProp as JsonSchema)?.items as JsonSchema)?.type !==
-									'object'))
+								(oneOfSchemaProp?.items as JsonSchema)?.type !== 'object'))
 					) {
-						_schemaProp = oneOfSchemaProp as JsonSchema;
+						_schemaProp = oneOfSchemaProp;
+						oneOfPath.push([key, index]);
 					}
-				}
+				});
 			}
 
 			if (isNullish(_schemaProp?.items)) {
@@ -235,6 +253,7 @@ export const convert = (
 						schema,
 						dataPath,
 						format,
+						oneOfPath,
 					);
 				}