+++ packages/web3-eth/CHANGELOG.md
@@ -56,3 +56,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Removed
 
 -   Moved the errors' classes from `web3-eth/src/errors.ts` to `web3-errors/src/errors/transaction_errors.ts` (#5462)
+
+### Fixed
+
+-   Fix `getBlock` returning empty transactions object on `hydrated` true (#5556)